+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * New `prompt()` and `prompt_file()` make it easier to create prompts that are a mix of static text and dynamic values.
 
+* You can find how many tokens you've used in the current session by calling `token_usage()`.
+
 * `chat_browser()` and `chat_console()` are now `live_browser()` and `live_console()`.
 
 * The `echo` can now be one of three values: "none", "text", or "all". If "all", you'll now see both user and assistant turns, and all content types will be printed, not just text. When running in the global environment, `echo` defaults to "text", and when running inside a function it defaults to "none".